requirements.txt
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==2.0.6
django-import-export==2.5.0
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
tablib==3.0.0
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
